+++ src/positionECEF.h
@@ -37,6 +37,7 @@
 #include "algorithm.h"
 #include "frameOfReference.h"
 #include "point.h"
+#include "ranges.h"
 #include <units.h>
 
 inline namespace coordinates
@@ -262,8 +263,8 @@ inline namespace coordinates
 		 * @returns		distance between this Point and <i>p</i> in units of distance_units.
 		 */
 		template<class Point>
-		[[nodiscard]] unit_type distance(const Point& p) const
-		{ return coordinates::distance(*this, p); }
+		[[nodiscard]] ranges::Euclidean distance(const Point& p) const
+		{ return ranges::Euclidean(coordinates::distance(*this, p)); }
 
 		/**
 		 * @brief		Calculates the dot product of two points.
@@ -280,7 +281,7 @@ inline namespace coordinates
 		 * @details
 		 * @returns		magnitude of the point vector wrt its origin.
 		 */
-		[[nodiscard]] unit_type magnitude() const { return coordinates::magnitude(*this); }
+		[[nodiscard]] ranges::Euclidean magnitude() const { return ranges::Euclidean(coordinates::magnitude(*this)); }
 
 		//////////////////////////////////////////////////////////////////////////
 		//		ACCESSORS