+++ src/positionENU.h
@@ -37,6 +37,7 @@
 #include "frameOfReference.h"
 #include "point.h"
 #include "positionGeodetic.h"
+#include "ranges.h"
 
 #include <cassert>
 #include <stdexcept>
@@ -375,8 +376,8 @@ inline namespace coordinates
 		 * @returns		distance between this Point and <i>p</i> in units of distance_units.
 		 */
 		template<is_point Point>
-		[[nodiscard]] distance_unit_type distance(const Point& p) const
-		{ return coordinates::distance(*this, p); }
+		[[nodiscard]] ranges::Euclidean distance(const Point& p) const
+		{ return ranges::Euclidean(coordinates::distance(*this, p)); }
 
 		/**
 		 * @brief		Calculates the dot product of two points.
@@ -393,7 +394,7 @@ inline namespace coordinates
 		 * @details
 		 * @returns		magnitude of the point vector wrt its origin.
 		 */
-		[[nodiscard]] distance_unit_type magnitude() const { return coordinates::magnitude(*this); }
+		[[nodiscard]] ranges::Euclidean magnitude() const { return ranges::Euclidean(coordinates::magnitude(*this)); }
 
 		//////////////////////////////////////////////////////////////////////////
 		//		ACCESSORS